requirements.txt
pydantic #0.18.2
pytorch #CUDA 12.1
tqdm
rotary-embedding-torch
einops